requirements.txt
keyboard==0.13.5
pkg-resources==0.0.0
pyserial==3.5
opencv-python>=4.5.0
numpy>=1.19.0
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
pillow>=8.3.0
werkzeug>=2.0.0
gunicorn>=20.1.0
